+++ src/kernel_evm/CHANGES.md
@@ -32,6 +32,7 @@
   returning them. (!10116)
 - Add support for `eth_getBlockTransactionCountByHash` and `eth_getBlockTransactionCountByNumber`. (!10101)
 - Add support for `web3_sha3`. (!10112)
+- Add support for `eth_getUncleCountByBlockHash` and `eth_getUncleCountByBlockNumber`. (!10131)
 
 
 ### Bug fixes